SFX-201 WIP: unit testing
