+++ packages/web-components/@sfx/product/karma.conf.js
@@ -1,0 +1,2 @@
+module.exports = require('../../../../karma.conf');
+
