+++ packages/web-components/@sfx/product/package.json
@@ -2,11 +2,13 @@
   "name": "@sfx/product",
   "version": "0.0.0",
   "description": "",
-  "main": "./src/index.ts",
+  "main": "dist/index.js",
+  "module": "src/index.ts",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
     "build": "../../../../scripts/build.sh",
-    "dev": "nodemon --watch src --exec npm run build"
+    "dev": "nodemon --config ../../../../nodemon.json --exec npm run build",
+    "tdd": "karma start --no-single-run --auto-watch",
+    "test": "karma start"
   },
   "repository": {
     "type": "git",
