+++ packages/web-components/@sfx/product/test/setup.ts
@@ -1,0 +1,1 @@
+import '../../../../../test-tools/setup';
