+++ packages/web-components/@sfx/product/test/utils.ts
@@ -1,0 +1,1 @@
+export * from '../../../../../test-tools/utils';
